Refactor code to get the selected compliance year along with unit tests
